feature(frontend): add client frontend

+++ backend/go_services/metadata/internal/domain/entities/genre.go
@@ -1,7 +1,7 @@
 package entities
 
 type Genre struct {
-	ID      int32     `gorm:"primaryKey" db:"id"`
+	ID      int64     `gorm:"primaryKey" db:"id"`
 	Name    string    `gorm:"column:name;not null;unique;size:25;" db:"name"`
 	Artists []*Artist `gorm:"many2many:artist_genres" db:"-"`
 	Albums  []*Album  `gorm:"many2many:album_genres" db:"-"`
